refactor: move timestamp/timerange formatter to formatting
